shill: passpoint: push EAP credentials to supplicant.
As Shill is not using wpa_supplicant interworking_connect function to
associate with a Passpoint compatible network, it is not providing the
EAP credentials to wpa_supplicant. Assuming FQDN, realm and OIs were
sufficient for network selection and match.

However to match with a network using the FQDN only, supplicant needs to
verify if the available EAP credentials are compatible with the EAP methods
advertised by the endpoint. Shill now pushes the EAP credentials when
adding a credentials block to supplicant to allow such matches.

BUG=b:225170348
TEST=FEATURES="test" emerge-${BOARD} shill
Create a Passpoint access point and provision the matching credentials
in Shill using the manager.AddCredentials() D-Bus call. Ensure Shill
matches and tries to connect to the access point.

Change-Id: Ib204101dabe395c3949468d0f0d6e7dd5e9bfafa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3532902
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>