cryptohome: Add optional value check in PinWeaverAuthBlock
When AuthState is set in VaultKeyset, it checks if these optional values
exist in vk or not and then sets it. However USS, which uses the same
AuthBlocks for key derivation doesn't include chaps key, So there is
a chance it may not be set so this CL accounts for that.
fek_iv and chaps_iv were set in AuthBlock for legacy vaultkeyset reasons.
This will change with USS as they will be handled outside.

BUG=b:222166546
TEST=existing unit tests pass and new unit tests as well
TEST=FEATURES=test emerge-zork cryptohome
TEST=tast run <ip> cryptohome.* hwsec.*

Change-Id: I925891ab5f6dd6dd1767510b766049e04de76b92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3533559
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>