crash: Ignore system crashes when no consent.
As part of the per-user metrics consent changes, it's possible that
there will be a system crash that occurs when a consenting user is
logged in, that crash is stored, and then subsequently the user logs out
before the crash can be uploaded. The user might then log in again,
and we want to upload the crash when they do.

BUG=b:214440403
TEST=unit tests; tast run forward_crbook crash.SenderNoConsent

Change-Id: Id8dd2f547523a6de02272d16856934d6a7853993
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3536625
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>