hps: Do not respawn hpsd after a failed update
Instruct upstart to not respawn hpsd on exit code 5. Exit with that code
when an update is sent, but the subsequent boot still requires an
update. This probably indicates a mismatch in signing keys, but could
have other causes. We do not want to continuously respawn hpsd and write
to the flash many times in this case.

BUG=b:205888733
TEST=start hpsd normally # normal boot
TEST=start hpsd with bad stage1 # hpsd starts only once, metric is
emitted, no crash is sent.

Change-Id: I973d3ec14a66d7ecf9f5c16236500f0ebe63f6f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3537002
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Dan Callaghan <dcallagh@chromium.org>
Tested-by: Evan Benn <evanbenn@chromium.org>
Commit-Queue: Evan Benn <evanbenn@chromium.org>