common-mk: Add oemcrypto_v16 USE flag
As we upgrade OEMCrypto versions over time on Intel, we will need to do
the FW upgrade in lockstep with the daemon build config in the OS image.
This will happen for volteer/brya in the coming months. In order to not
require that both brya and volteer FW for all models are updated at the
same time this flag is needed so we can differentiate between volteer
and brya. We will then also use this flag for determining OEMCrypto
version on other platforms as well then.

BUG=b:185144739
TEST=None

Change-Id: I757d3a2c11c1b5f84f1158126d371811dc15aaf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3541085
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>