arc-setup: Handle the fact that machine SoC info ID isn't always soc0
Currently to find the machine on ARM arc-setup hardcodes to look in
`/sys/devices/soc0/machine`. The problem is that the ID number "0" is
actually dynamic and there may be more than one SoC device. I've
booted a kernel and seen that the `soc0` is actually populated by a
different driver (the one providing jep106 info).

Fix it so that:
* We enumerate over _all_ SoC drivers (everything under the "soc"
  bus).
* We keep searching until we find a match.

NOTE: this is what things look like my trogdor running Chrome OS
kernel 5.4:
  $ grep "" /sys/bus/soc/devices/*/{family,machine}
  /sys/bus/soc/devices/soc0/family:Snapdragon
  /sys/bus/soc/devices/soc1/family:jep106:0070
  /sys/bus/soc/devices/soc0/machine:SC7180

On a different Qualcomm board running a different kernel:
  $ grep "" /sys/bus/soc/devices/*/{family,machine}
  /sys/devices/soc0/family:jep106:0070
  /sys/devices/soc1/family:Snapdragon
  /sys/devices/soc1/machine:SC7280

Note that the swapping of soc0 for soc1 isn't something intentional
but it just happenstance because those numbers aren't guaranteed:

As part of this we switch _away_ from the "safe" version of the file
reading function. This is on purpose. Linux sysfs by design has lots
of symlinks and the "safe" versions of the functions block these. For
instance:
  $  ls -l /sys/bus/soc/devices
  total 0
  lrwxrwxrwx. 1 root root 0 Mar 25 15:24 soc0 -> ../../../devices/soc0
  lrwxrwxrwx. 1 root root 0 Mar 25 15:24 soc1 -> ../../../devices/soc1

Userspace can't write to sysfs anyway so we shouldn't need to worry
about someone doing a symlink attack there and there's a benefit to
not needing to hardcode a naming syntax for soc devices.

BUG=b:223410173, b:175610620
TEST=`android-sh -c getprop | grep -w soc` on ARM
TEST=`FEATURES=test emerge-dedede arc-setup`

Change-Id: I2d4c5c39c505ad91e2b4597c932197c76f7e3c00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3541097
Tested-by: Douglas Anderson <dianders@chromium.org>
Feels: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>