hps: Add post-stop sleep to slow down respawn
We have observed the i2c bus getting stuck in a bad state sometimes. The
i2c bus remains in the bad state while hps is constantly trying to
communicate with it. The i2c bus recovers if left alone for a few
seconds to reset itself.

https://source.chromium.org/chromiumos/chromiumos/codesearch/+/HEAD:src/third_party/kernel/v5.10/drivers/i2c/busses/i2c-designware-master.c;l=496;drc=eaa42770c5cd35a65b7d1554e0c5973cb307b293

BUG=b:205888733, b:198515585, b:205659252
TEST=stop hpsd; # quick, no sleep
TEST=restart hpsd; # slow, does the sleep
TEST=start hpsd with crash bug # slow, does the sleep

Change-Id: I139f079ef52924ac1c8882d498cffe7f6aacd23c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3547319
Reviewed-by: Dan Callaghan <dcallagh@chromium.org>
Commit-Queue: Evan Benn <evanbenn@chromium.org>
Tested-by: Evan Benn <evanbenn@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>