chromeos-dbus-bindings: add method helper for adaptor generator
This CL adds a part of adaptor generator helpers. The added helpers make
a return type and arguments and are used by a tmplate to generate
interface methods. The function contents in the signature package will
be implemented in the following CLs.

This is a part of a work to rewrite chromeos-dbus-bindings generator in Go.

BUG=chromium:983008
TEST=FEATURES=test emerge-<board> chromeos-base/chromeos-dbus-bindings

Change-Id: I87c94d7df7d1cc15fe85dffd7a39e7fc142940b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3548316
Tested-by: Mai Yokozeki <yokozekimai@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Mai Yokozeki <yokozekimai@google.com>