rmad: Fix calibration error handling
In the original design, we didn't have a different path to get the next
state case for normal use and startup. So we implemented relative error
handling in the normal path (InitializeState + GetNextStateCase).
However, now that we've separated it, we should move the relevant logic
into the correct function, otherwise the error handling will fail.

BUG=b:218466072
TEST=FEATURES=test emerge-octopus rmad

Change-Id: Id336f2cb414ae5b7eee1bff3ea4c3062a3e9f061
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3552139
Tested-by: Gene Chang <genechang@google.com>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>
Commit-Queue: Gene Chang <genechang@google.com>