shill: report adapter HW info
Report adapter information in a different structured metric project to
avoid correlation with actual connection data.
Adapter information collected with connection data will go through an
allowlist (still to be implemented) based on popularity of the adapters
and/or values that are to be expected based on AVL. For now we report
the default error value in that case.

BUG=b:227974274
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill

Change-Id: Ib1eda52d1da6058f9a9da2d987467de18e41a53a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3567366
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>