arc-setup: Removing Realpath() in mount call
The current implementation of Mount() function
calls Realpath() in source and targets before making the
mount() call. This leads to symlinks being resolved
and poses risk for mounting unintended locations.

Based on the results of this test CL -
https://crrev.com/c/3539286, there were no
differences observed between source and target,
and their Realpath() counterparts.

Hence, with this CL -

1) Removing the Realpath() calls on
source and target, and passing them to mount() directly.

2) Removing BindMountWithNoPathResolution() function
and any calls to it.

BUG=b:204222140
TEST=Manual, no mounting error logs observed in
/var/log/arc.log

Change-Id: I41bbca5a520167780b47b648fd9f19336e8e274d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3552673
Reviewed-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>