shill: vpn: VPNService reads technology from Service
The VPNService class needs to read the technology type of the underlying
physical service. Currently this is done by read that property of the
Connection class, and this patch changes it to read the value from
Service. Note that the technology type of Connection and Service might
be different in some cases (e.g., EthernetEAP and Cellular). According to
the description in service-api.txt, the technology of service is
actually expected, and technically correct for the currenct usage in
update_engine for the cellular cases.

After this patch, VPNService class no longer needs to access the
Connection class directly.

BUG=b:182777518
TEST=unit tests

Change-Id: I076f13f90c2f8d8518dab3033d5eeb2071704169
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3552629
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>