vm_tools: Parse VVU devices' socket indices
This change is relevant to manatee. We now read each VVU device's UUID
and parse the socket index from it. This is then passed to the
corresponding sibling driver in the sibling VM as the socket indice to
construct the path of VVU protocol socket.

BUG=b:196186396
BUG=b:215472603
TEST="vmc start termina" on manatee boards.

Cq-Depend: chromium:3488307
Change-Id: If0ea0e70f994271cf46baf22df6d5c198b3054df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3557169
Tested-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Auto-Submit: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>