u2fd: Start using string request_id
Start calling the new ShowAuthDialogV2 method that takes a string as
request id. If the request we received from Chrome contains a non-empty
request_id_str we use that string, otherwise we stringify the integer
request_id field. Similar logic applies to where we compare whether the
request id of the Cancel request is same as the pending session's.

BUG=chromium:1269528
TEST=u2f unittests
TEST=tast run $DUT u2fd.*

Change-Id: Ife6a99a6c6af3fe493f8a7eac51568ddd86d413f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3555513
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>