debugd: Specify encoding for drm_trace in LogTool
In the case that some non-UTF8 character ends up in the trace, we want
to strip that out of the log and leave the remainder to be human
readable. This is accomplished by specifying the encoding. Leaving it as
kAutoDetect would base64 encode the logs if a non-UTF8 character was
found.

BUG=b:222539380
TEST=dbus-send org.chromium.debugd.GetLog "string:drm_trace" on copano

Change-Id: Ia6c77cb69fabcc756cf41cba7ce3e7c3bdd6a190
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3558050
Tested-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>