ml: add access to seccomp for hwr
When resolving the models paths for a LoadHandwritingModel request that
contains a spec->language_pack_path, we use `bath::PathExists` to verify
the path exists. This method then calls `access` which is blocked by the
seccomp, thus crashing the ml_service.

This CL updates the seccomp for handwriting to include `access`.

BUG=1306350, b:227423469
TEST=cros_run_unit_tests --board=${BOARD} --packages chromeos-base/ml

Change-Id: I76a5343916e99119c34cfb9482335ddef1788f5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3562400
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: Claudio M <claudiomagni@chromium.org>
Tested-by: Claudio M <claudiomagni@chromium.org>
Commit-Queue: Curtis McMullan <curtismcmullan@chromium.org>