shill: metrics: remove DarkResume and WakeOnWiFi metrics
The DarkResume / WakeOnWiFi feature is not functional in its current
implementation and has been disabled for a while. The histograms
associated with legacy metrics added in 2016 are either expired or empty:
  https://uma.googleplex.com/p/chrome/timeline_v2?sid=d5343122bd4b80b0cd9e952a43fab424

Since the implementation of these legacy metrics are likely to be
incorrect or obsolete after the WakeOnWiFi feature is revived, this
patch proposes to remove all the associated metrics in order to simplify
shill code and reduce metrics maintenance.

Note that all methods defined in WakeOnWiFiInterface are preserved even
if multiple methods are now noops in WakeOnWiFi.

BUG=b:172212984
BUG=chromium:1261033
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I177a9a20871f8ae0eab9e69bbfb6353a997190e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3565062
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>