vtpm: prepare dbus interface definition.
We are about to build the dbus service for vtpm. Because of its nature,
it is a trunks-like service, so this CL prepares the future DBus
definition by copying the definitions from trunks w/ proper rename of
the filename as well as the content.

BUG=b:227341806
TEST=CQ.

Change-Id: I96abf17d894e33e8993f71146c9e43375df480c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3563663
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>