vtpm: delegate the the command execution to `Command` object.
The effort is made so we can decouple the D-Bus service model and the
vTPM core functoins, which we will fully focus on.

In more details, now the service delegates the vTPM behavior to
`Command`. Also, it take care of the threading issue so any `Command`
object doesn't have to be worried about the location of the D-Bus
response being `Return()`ed.

BUG=b:228789530
TEST=unittests
TEST=`trunks_client --vtpm --tpm_version` gets expected error
      w/ CL:3563664 (TPM_RC_INSUFFICIENT)

Cq-Depend: chromium:3578950
Change-Id: I003a90a0a4cdd39f30f0b78695ba4986ac5a42de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3578102
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>