trunks: dbus proxy support trunks-like dbus interfaces.
We are about to have vtpm service, of which the dbus methods are exactly
the same as trunks' because its nature. To adopt vtpm as the first stop
of handling tpm requests from consumers, it is an efficient way to share
codes so vtpm package doesn't have to maintain its own client for dev
purpose.

BUG=b:227471635
TEST=`trunks_client --tpm_version` works correctly.

Change-Id: I538236b4c13658022fbfe2665ef021e2359f043a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3567203
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>