crostini_client: switch `vmc create` to use getopt
The existing cli parsing for `create` is awkward to extend because of
its use of `--` as a separator. By setting StopAtFirstFree, though,
we're able to maintain the current interface for additional params while
allowing the command to be easily extended in the future.

TEST=deployed to hatch device, ensured parameters continue to be parsed
correctly. These are also coverd by test cases.

BUG=b:215756557

Change-Id: I0c3792137261ad3a3116810b25a77d4e8aab0365
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3563686
Tested-by: Mike Gerow <gerow@google.com>
Auto-Submit: Mike Gerow <gerow@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Mike Gerow <gerow@google.com>