hps: add support for booting stage0 v4
Corresponds to this firmware change: https://crrev.com/c/3561248

The AVERIFY / ANONVER status bits are removed. Stage0 no longer attempts
to validate stage1 on reset. Instead, when the launch1 command is sent,
stage0 validates stage1 and jumps immediately to it, or sets an error
bit if stage1 is not valid.

We can examine the "hardware version" register to check whether we're
talking to stage0 v3 or v4. For older versions, we preserve the existing
boot flow. For v4 and later, we instead just send the launch1 command
immediately and then check for errors or version mismatches afterwards.

BUG=b:226542882
TEST=start hpsd ; restart hpsd  # with stage0 v3
TEST=start hpsd ; restart hpsd  # with stage0 v4

Change-Id: Ief8ca88827d01af848bcb2078bbe40c586a841c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3565066
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>