secanomalyd: Introduce the concept of "known" mounts.
/run/arc/shared_mounts/data is an example of a mount that is currently
reporting as W+X, but is understood, we're in the process of fixing,
and is low-risk. In this case the mount is low-risk because it only
exists on the login screen but not in a logged-in user session.

Moreover, the relative frequency of /run/arc/shared_mounts/data is
forcing us to sample reports and could be drowning out less frequent
(but likely more interesting) anomalous mounts.

Associate each known mount with a bug tracking its fix but also skip
reporting it since they're polluting our numbers.

BUG=b:227800972
TEST=New unit tests.
TEST=Deploy to device, run on device at the login screen,
TEST=see mount skipped.

Change-Id: Ic98c6e6c9946eeba13c7b935bc00ba04868452b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3563690
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>