system_api: Add powerd SuspendImminent.action
Add an optional action member to the SuspendImminent signal. Consumers
of SuspendImminent can use this to figure out if the system is going
down for a suspend, or about to transition through a hibernate resume.
Usually, the same behavior is appropriate for both actions. Consumers of
this signal only need to become aware of this new member if they need to
do something different before resuming from hibernate than they would
for a suspend.

BUG=b:219779783
TEST=emerge-$BOARD system_api power_manager

Change-Id: I40f55ccd81a89c048d55cd37b1762ca8a65476a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3565967
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>