shill: Include VPN devices in Manager::FindDeviceFromService()
The current implementation of Manager::FindDeviceFromService() iterates
over |devices_| which does not include virtual devices, and thus it will
return a nullptr for a VPNService. This breaks the logic that find the
Connection associated with a Service using this function which was added
in https://crrev.com/c/3559092 . This patch modifies this behavior to
include the VPN devices.

Currently this function is used in:
- Manager::FindConnectionFromService(): the modified behavior is
  expected;
- Manager::SortServicesTask(): this change has no effect here;
- Service::RequestTrafficCounters(): the modified behavior should be
  expected to get the interface name of a VPN service.

BUG=b:227964549
TEST=unittests
TEST=tast run $DUT network.VPNConnect.wireguard
TEST=manually checked the connectivity of a VPN service

Change-Id: I0aeb844fe57b1b13b0e27b3561bfea379af09a94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3563817
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>