dlcservice: Fail on early installations
There is a race where dlcservice sends an installation request, but
update_engine returns a signal indicating that an installation isn't in
progress. This causes dlcservice to terminate/cancel the current
installation. To disambiguate this race, return a failure when
update_engine service isn't up from dlcservice point of view and an
installation is started.

BUG=b:227807674
TEST=emerge-$B dlcservice

Change-Id: I342bbfd295b2bb6ed051355a3ee26f0e740c24d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3573104
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Henry Barnor <hbarnor@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>