shill: Allow switching slots even if primary SIM is scanning.
On dual SIM MBIM platforms, the eSIM is exported as a SIM object.
https://gitlab.freedesktop.org/mobile-broadband/ModemManager/-/commit/9940bd9dbe88430318ae8f045001db68183b42c5
Cellular is in a perpetual scanning state if the eSIM does not have any
profiles (since the modem is in enabled state). Shill should allow
switching slots even if the current SIM is still scanning.

This problem does not occur on other platforms since a profileless
eSIM will put the modem in failed state and thus Shill does not
set scanning=true.

The issue of shill being in a perpetual scanning state when the eSIM is
profileless also needs to be fixed but that does not have functional
implications.

BUG=b:212210640
TEST=tast run ${cr} '("group:cellular" && "cellular_sim_active")' on all
modems.

Change-Id: I8f584a4b02caedfe65ab285de4d692394774c7d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3569391
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Pavan Holla <pholla@google.com>
Commit-Queue: Pavan Holla <pholla@google.com>