shill: vpn: Avoid pppd logging to stdout
By default, pppd will send log messages to stdout, unless fd=1 is used
by the actual data channel (which is the case of the new swanctl-based
driver for L2TP/IPsec VPN), but this checking code happens after all
options are processed, and logging may happen when processing the plugin
option. As a result, pppd will send this log message to the peer, and if
peer rejects but not ignores an invalid ppp packet, the connection will
be closed.

This patch adds the option in the config file to avoid pppd logging to
stdout.

BUG=b:218437737
TEST=unit_tests;
TEST=tast run $DUT network.VPNConnect.l2tp_ipsec*;
TEST=manually checked the connectivity to the VPN provided in BUG.

Change-Id: Iab2a310703a13f94903f4e652ab75ed4306a026f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3569930
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>