mount-passthrough: Fix write support of ARCVM "Play files"
Currently, in ARCVM, when a regular file is created in "Play files" from
the host (Chrome OS) side, the file is not writable by Android apps.
This is because the group permission is set read-only in
passthrough_create().
This CL fixes this issue by adding the write bit to the group permission
when mount-passthrough is running for the ARCVM "Play files" sharing.
With this change, newly created files (including those that are copied
from another location with the Files app) will have the same permission
(media_rw:media_rw / 0664) as those created in ARC++ container.
One limitation is that files already existing as read-only will not be
writable unless it is newly copied to another location.

BUG=b:228123869
TEST=Launch the Files app. Copy a text file from "Google Drive" to "Play
files" > Documents. Open, edit, and save the file with an Android app.

Change-Id: I2726b719f262017edbd222a7f3f40a83c9af363d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3573201
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Risan <risan@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>