cryptohome: Handle hibernation resume device
During hibernation, the stateful partition is set up on top of
dm-snapshot. We use the presence of the device as an indicator
that system is currently in resume-from-hibernation and that
all logical volume related operations should use the dm-snapshot
device instead of the stateful partition.

BUG=none
TEST=CQ

Change-Id: I63c00165454596945bf8840bdc5052e104bbf855
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3573105
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>