hiberman: Tolerate fiemap shrinkage
I'm seeing instances where the kernel will report 7 fiemap entries when
no buffer is supplied, but then only 6 when a buffer is actually
supplied. This is happening even though I know nobody else is changing
the hibernate data file.

The original error was meant to catch someone else modifying the file
size while hiberman was using it. Since that's not the case, tolerate
shrinkage from what was reported before the buffer to what was reported
after. If the fiemap grows between the calls, that is still a problem
that results in failure.

BUG=b:204563517
TEST=hiberman resume --test-keys on brya

Change-Id: I94d7108aa179eac3660afa4bb118588e39f504a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3576578
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>