hiberman: Zero fill DiskFiles on LVM-enabled systems
On systems where LVM is enabled, calling fallocate() is not enough. The
thin logical volume disk extents backing that logical disk space may not
actually be allocated. So at resume, if we were to go access those disk
regions, those blocks would not actually be backed by anything. The
free/busy bitmaps in the thinpool metadata area would change, and after
resume the kernel would be faced with corruption.

There currently is no equivalent of fallocate() for thin volumes, other
than to actually touch the disk areas you want to use. So for now, zero
out the hibernate file disk extents at the beginning of hibernate time.
This could end up being a fairly expensive operation in terms of time
and disk I/O. Fortunately it should be a one-time cost, as once the
hibernate files are set up, they don't get deleted unless the user is
low on disk space.

Once we get a primitive where we can allocate that disk space from the
thinpool directly, we'll switch to using that and stop zeroing out the
files here.

BUG=b:202277660
TEST=hibernate on brya

Change-Id: I62a332050ab06803d489185e32b8b333357e4843
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3576579
Reviewed-by: Alexis Savery <asavery@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>