shill: Move DHCP callbacks from IPConfig to DHCPConfig (part 1)
The callbacks in IPConfig (update, failure, and expire) are actually
only for DHCP and triggered from DHCPConfig class, and thus they can and
should be moved into DHCPConfig. This patch breaks the inheritance
relationship of IPConfig::UpdateProperties(), and thus can
NotifyUpdate() can be moved into DHCPConfig easily in the following
steps.

The only possible logic change in this patch is, we used to call
EmitChanges() after invoke update callback, but after this patch this
order is changed. This should not have any real impact since
EmitChanges() will be called anywhere after any property is changed in
IPConfig, so even if the callback updates any property, EmitChanges()
will be called again and thus the listener can get the event.

Also uses mockable keyword in ipconfig.h to make it clear that which
functions can be overridden.

BUG=b:227560694
TEST=unit tests

Change-Id: Iff297ad03f3ad9bb877f72bae1a0eaaa75e44a8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3572470
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>