resourced: Update D-Bus service to use PowerPreferencesManager
This will enable resourced to start consuming and applying the
chromeos-config preferences. This will be a no-op for all platforms that
don't define the resource chromeos-config values.

We don't currently have a listener on the `PowerSupplyPoll` message, so
the power preferences will only be applied when one of the activity
indicators change. This will be added in a followup.

BUG=b:221091499
TEST=Manually tested on guybrush and verify powersave_bias is updated
correctly.

Cq-Depend: chromium:3642997
Change-Id: I1308348a64da1bc27e467530e71fee3756008334
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3575557
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Vovo Yang <vovoy@chromium.org>