shill: allowlist for WiFi adapter info
Populate an allowlist of known hardware adapters that have been added to
AVL. If the probed IDs are in the allowlist, then we can report them.
For now the allowlist is the list of WiFi chipsets added to AVL.
Depending on the popularity of adapters not currently in AVL, we could
later create a second allowlist that would include adapters that are
sufficiently popular.

BUG=b:227974274
TEST=cros-workon-dedede start shill
TEST=FEATURES=test emerge-dedede shill
TEST=emerge-dedede shill
TEST=cros deploy $DUT shill
TEST=check chrome://histograms#Network.Shill.WiFi.AdapterAllowlisted
TEST=repeat with AX201 removed from AVL list, and check
chrome://histograms#Network.Shill.WiFi.AdapterAllowlisted again

Change-Id: I8fdbb9a43a61d25a48f17e960ceb75567a1661da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3582962
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: David Ruth <druth@google.com>
Reviewed-by: Shreyas Karkhedkar <karkhedkar@google.com>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>