power: Add and build Adaptive Charging Files
This adds the new AdaptiveChargingController class, plus other directly
related functionality, but doesn't use it yet (which will be done in the
next CL). This is for the Adaptive Charging feature, which will delay
charging as long as we predict that the AC charger will remain plugged
in.

BUG=chromium:1243163
TEST=cros_run_unit_tests --board atlas --packages "power_manager"

Change-Id: I388c0591d3fb34bc932b6d748e236e24b00d2ad9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3575561
Tested-by: Derek Basehore <dbasehore@chromium.org>
Auto-Submit: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>