cros-disks: Simplify Process class
When capturing the messages written by a subprocess, cros-disks does not
need to distinguish between messages sent to stdout and messages sent to
stderr. Therefore, the Process class now connects both stdout and stderr
to the same pipe, which simplifies the polling and the reading of this
pipe.

BUG=chromium:1158662
TEST=cros_workon_make cros-disks --test

Change-Id: Ie615182e9b37a8e73bf7c5e322cd22e4985977bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3577760
Reviewed-by: Marcello Salomao <msalomao@google.com>
Commit-Queue: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>