hermes: Delay inhibit if modem is in a transitional state
On Brya, MM inhibit returns a DBus error if the modem state is "enabling".
MM proceeds to inhibit the modem after reaching enabled. This leads to
the system being unusable because Hermes assumes the inhibit failed
whereas MM actually inhibits the modem. We still need to root cause
the DBus failure, this CL is just a workaround.

BUG=b:228314490
TEST=Flash a new image, check that MM is inhibited successfully by Hermes

Change-Id: I614aea278b7faacee47585e1c8945a96492f052e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3578014
Tested-by: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Pavan Holla <pholla@google.com>