libipp: Refactoring - new class Frame
This class is a part of new API. It will replace the current API
consisting of classes Client, Server, Respond and Request. The goals
are: making the API simpler, increase the test coverage, and decrease
the size of the library. Missing test for the parser and the method
SaveToBuffer(...) will be added in other CL.

BUG=none
TEST=FEATURES=test emerge-atlas libipp

Change-Id: I2413fe1760863d3788b78c4c4120e095e9d4e535
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3577696
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>
Auto-Submit: Piotr Pawliczek <pawliczek@chromium.org>
Commit-Queue: Piotr Pawliczek <pawliczek@chromium.org>