shill: Split Device::OnIPConfigUpdated()
This patch splits Device::OnIPConfigUpdated() into two functions:
OnIPConfigUpdatedFromDHCP() and OnIPConfigUpdated(). The former one is
used as the callback registered with DHCPConfig and gets called when we
get a new DHCP lease, while the latter one can be called from several
places indicates that IPConfig has been updated.

The only logic change in this patch is, WiFi class used to overrides
this function to perform the WakeOnWiFi logic. Except for the DHCP
callback, there are 4 call sites of this function in the current code:
- Device::AssignIPConfig();
- Device::AssignIPv6Config();
- VirtualDevice::UpdateIPConfig();
- Device::ConfigureStaticIPTask().
The first three are only used by Cellular and VPN and thus will not
affect the WiFi logic. According to the comments in wake_on_wifi.h,
WiFi only cares about DHCP but not static IP config changes, so this
change should be expected.

This patch is required to change the signature of DHCP callback.

BUG=b:227560694
TEST=unit test

Change-Id: I875e8dca6ec68d12e54db6feb44076c2f9d6ac1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3577769
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>