debugd: Replace property CrashSenderTestMode by method.
Current CrashSenderTestMode is implemented by a property.
However, in the current way, the signal propergation is blocked.
Also, this does not match the expected use of chromeos-dbus-bindings-generator.

Looking at the usage, it is only set just by a Tast test lib.
So, replace it by a method.

BUG=chromium:983008
TEST=Ran on DUT and made sure SetCrashSenderTestMode is actually used.

Cq-Depend: chromium:3578601
Change-Id: I52a40e8198a4f3a206a1d6d0329d7e642f73775a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3579067
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>