secanomalyd: Introduce the concept of "context checking".
In a previous CL we introduced the concept of a "known" mount. These
are W+X mounts that we know exist, we know are not immediately
dangerous, and we're tracking a fix for. Uploading these known mounts,
which are usually frequent, drowns out other signals.

In order to prevent these mounts from becoming blessed places to keep
W+X mounts around, introduce the concept of "context checking". Instead
of unconditionally skipping known mounts, make sure that the mount is
only skipped if it's happening in its expected context. For example, a
mount that's deemed low risk because it only occurs on the login screen
should only be skipped when it occurs on the login screen.

This CL introduces a SystemContext class that will contain all the
signals required to check the context of a mount. The SystemContext
class is however not yet doing anything -- adding its functionality
will come in a later CL. Because the only existing signal ("is the
system logged in?") is hardcoded to 'false', this CL does not change
uploading behaviour: before this CL, and with this CL,
/run/arc/shared_mounts/data is skipped.

The CL also introduces a SystemContextMock class used for unit tests.
The mock constructor forces the caller to specify all values for
signals to make tests clearer. The Refresh() function is also mocked
even though it's not currently being called, for completeness.

BUG=b:228638237
TEST=New unit tests.
TEST=Deploy to device, run secanomalyd on the login screen, mount is
TEST=skipped.

Change-Id: Iba584ba1c3326afb1c6da3ff21a1fd5d6fbd7401
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3577477
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>